+++ libzet/__init__.py
@@ -6,3 +6,8 @@ else:
     import importlib_metadata as metadata
 
 __version__ = metadata.version('libzet')
+
+from libzet.Zettel import (
+    load_zettels, save_zettels, SkipZettel, str_to_zettels, Zettel, zettels_to_str)
+
+from libzet.edit import create_zettel, edit_zettels