+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+[unreleased]
+============
+
+Added
+-----
+- Can import important functions directly from libzet.
+  eg. ``from libzet import Zettel``
+
 [1.0.0-alpha] - 2023-03-08
 ==========================
 New general flow for using a Zettel.